refactor (Money): move Money to own class.
