+++ Fluid/Filters/MiscFilters.cs
@@ -33,10 +33,6 @@ namespace Fluid.Filters
             filters.AddFilter("escape_once", EscapeOnce);
             filters.AddFilter("handle", Handleize);
             filters.AddFilter("handleize", Handleize);
-            filters.AddFilter("money", Money);
-            filters.AddFilter("money_with_currency", MoneyWithCurrency);
-            filters.AddFilter("money_without_currency", MoneyWithOutCurrency);
-            filters.AddFilter("money_without_currency", MoneyWithoutTrailingZeros);
 
             return filters;
         }
@@ -124,52 +120,6 @@ namespace Fluid.Filters
         public static FluidValue EscapeOnce(FluidValue input, FilterArguments arguments, TemplateContext context)
         {
             return new StringValue(WebUtility.HtmlEncode(WebUtility.HtmlDecode(input.ToStringValue())));
-        }
-
-        /// <summary>
-        /// Alias for Money Without Currency
-        /// </summary>
-        public static FluidValue Money(FluidValue input, FilterArguments arguments, TemplateContext context)
-        {
-            return MoneyWithOutCurrency(input, arguments, context);
-        }
-
-        public static FluidValue MoneyWithCurrency(FluidValue input, FilterArguments arguments, TemplateContext context)
-        {
-            if (!decimal.TryParse(input.ToStringValue(), out decimal result))
-            {
-                return NilValue.Instance;
-            }
-
-            return new StringValue(result.ToString("C", context.CultureInfo));
-        }
-
-        public static FluidValue MoneyWithoutTrailingZeros(FluidValue input, FilterArguments arguments, TemplateContext context)
-        {
-            if (!decimal.TryParse(input.ToStringValue(), out decimal result))
-            {
-                return NilValue.Instance;
-            }
-
-            var currencyValue = result.ToString("C", context.CultureInfo);
-            if (currencyValue.Contains(context.CultureInfo.NumberFormat.CurrencyDecimalSeparator))
-            {
-                currencyValue = currencyValue.TrimEnd('0');
-                if (currencyValue.EndsWith(context.CultureInfo.NumberFormat.CurrencyDecimalSeparator))
-                    currencyValue = currencyValue.Replace(context.CultureInfo.NumberFormat.CurrencyDecimalSeparator, "");
-            }
-            return new StringValue(currencyValue);
-
-        }
-
-
-        public static FluidValue MoneyWithOutCurrency(FluidValue input, FilterArguments arguments, TemplateContext context)
-        {
-            if (!decimal.TryParse(input.ToStringValue(), out decimal result))
-            {
-                return NilValue.Instance;
-            }
-            return new StringValue(result.ToString("C", context.CultureInfo).Replace(context.CultureInfo.NumberFormat.CurrencySymbol,""));
         }
 
         public static FluidValue Date(FluidValue input, FilterArguments arguments, TemplateContext context)
